requirements.txt
langchain==0.0.278
langchain_experimental==0.0.12
python-dotenv==1.0.0
streamlit==1.26.0
